Added full CNN sample, bug fixes, some improvements
- Added example of digit recognition with touchscreen for Arduino Mega.
- Same bugs fixed.
- better implementation of fixed exponential function for fixed8 and fixed16

+++ embedia/libraries/float/embedia.c
@@ -398,7 +398,7 @@ void softplus_activation(float *data, uint32_t length){
  *  *output => pointer to the data1d_t structure where the result will be stored.
  */
  void flatten3d_layer(data3d_t input, data1d_t * output){
-    uint16_t c,i,j;
+    uint32_t c,i,j;
     uint32_t cantidad = 0;
 
     output->length = input.channels * input.height * input.width;
@@ -503,7 +503,7 @@ void batch_normalization3d_layer(batch_normalization_layer_t layer, data3d_t *da
  */
 void image_adapt_layer(data3d_t input, data3d_t * output){
 
-    int i, j, c, l;
+    uint32_t i, j, c, l;
 
     output->channels = input.channels;
     output->height   = input.height;
